refactor: extract legacy/compat + core health endpoints from orchestrator (#548)
Phase 4.3 of orchestrator decomposition (Strangler Fig):
- Extract 5 legacy telephony endpoints (/tts, /stt, /chat, /process_call,
  /reset_conversation) + 5 OpenAI-compatible endpoints (/v1/*) to
  modules/compat/router.py (~510 lines)
- Extract core endpoints (/, /health, /admin/deployment-mode) to
  modules/core/router_health.py (~110 lines)
- Remove dead duplicate auth endpoints and /admin/status (shadowed by routers)
- All globals replaced with ServiceContainer via get_container()
- Orchestrator: 3544 → 2875 lines (-669)
