+++ orchestrator.py
@@ -5,16 +5,13 @@ STT (Whisper) -> LLM (vLLM / Cloud) -> TTS (XTTS v2)
 """
 
 import asyncio
-import json
 import logging
 import os
-import threading
 import time
 from datetime import datetime, timedelta
 from pathlib import Path
 from typing import Dict, List, Optional
 
-import soundfile as sf
 import uvicorn
 from fastapi import Depends, FastAPI, File, HTTPException, Request, UploadFile
 from fastapi.responses import (
@@ -66,12 +63,7 @@ from app.security_headers import (
     SecurityHeadersMiddleware,
 )
 from auth_manager import (
-    LoginRequest,
-    LoginResponse,
     User,
-    authenticate_user,
-    create_session,
-    get_auth_status,
     get_current_user,
     require_permission,
 )
@@ -86,7 +78,6 @@ from db.integration import (
     async_role_manager,
     async_user_manager,
     async_workspace_manager,
-    get_database_status,
     init_database,
     shutdown_database,
 )
@@ -222,6 +213,14 @@ app.include_router(claude_code.router)
 app.include_router(kanban.router)
 app.include_router(roles.router)
 app.include_router(workspace.router)
+
+# Core health + legacy/compat endpoints (Phase 4.3)
+from modules.compat.router import router as compat_router  # noqa: E402
+from modules.core.router_health import router as health_router  # noqa: E402
+
+
+app.include_router(health_router)
+app.include_router(compat_router)
 
 # Hardware/GPU routers — skip in cloud mode
 if DEPLOYMENT_MODE != "cloud":
@@ -549,41 +548,6 @@ async def _seed_default_workspace():
             logger.info(f"🏢 Workspace: ensured {added} users in default workspace")
     except Exception as e:
         logger.error(f"🏢 Workspace seed failed: {e}")
-
-
-class ConversationRequest(BaseModel):
-    text: str
-    session_id: Optional[str] = None
-
-
-class TTSRequest(BaseModel):
-    text: str
-    language: str = "ru"
-
-
-class OpenAISpeechRequest(BaseModel):
-    """OpenAI-compatible TTS request for OpenWebUI integration"""
-
-    model: str = "marina-voice"
-    input: str
-    voice: str = "marina"
-    response_format: str = "wav"
-    speed: float = 1.0
-
-
-class ChatMessage(BaseModel):
-    role: str
-    content: str
-
-
-class ChatCompletionRequest(BaseModel):
-    """OpenAI-compatible chat completion request"""
-
-    model: str = "anna-secretary-qwen"  # Format: {persona}-secretary-{backend}
-    messages: List[ChatMessage]
-    stream: bool = False
-    temperature: Optional[float] = None
-    max_tokens: Optional[int] = None
 
 
 @app.on_event("startup")
@@ -1093,569 +1057,11 @@ async def shutdown_event():
     logger.info("✅ Shutdown complete")
 
 
-@app.get("/")
-async def root():
-    """Проверка работоспособности"""
-    return {
-        "status": "ok",
-        "service": "AI Secretary Orchestrator",
-        "endpoints": {
-            "health": "/health",
-            "process_call": "/process_call (POST)",
-            "tts": "/tts (POST)",
-            "stt": "/stt (POST)",
-            "chat": "/chat (POST)",
-        },
-    }
-
-
-@app.get("/health")
-async def health_check():
-    """Проверка здоровья всех сервисов"""
-    # Используем llm_service из container (может быть обновлён через router)
-    from app.dependencies import get_container
-
-    container = get_container()
-    current_llm_service = container.llm_service if container.llm_service else llm_service
-
-    # Определяем тип LLM сервиса
-    llm_backend_type = "unknown"
-    if current_llm_service:
-        if isinstance(current_llm_service, CloudLLMService):
-            ptype = getattr(current_llm_service, "provider_type", "cloud")
-            llm_backend_type = f"cloud ({ptype}: {current_llm_service.model_name})"
-        elif hasattr(current_llm_service, "api_url"):  # vLLM
-            llm_backend_type = f"vllm ({current_llm_service.model_name})"
-        elif hasattr(current_llm_service, "model_name"):
-            llm_backend_type = f"cloud ({current_llm_service.model_name})"
-
-    services_status = {
-        "voice_clone_xtts_anna": anna_voice_service is not None,
-        "voice_clone_xtts_marina": voice_service is not None,
-        "voice_clone_openvoice": openvoice_service is not None,
-        "piper_tts": piper_service is not None,
-        "stt": stt_service is not None,
-        "llm": current_llm_service is not None,
-        "llm_backend": llm_backend_type,
-        "streaming_tts": streaming_tts_manager is not None,
-        "current_voice": current_voice_config,
-    }
-
-    # Для health check достаточно любой TTS + llm
-    any_tts = (
-        services_status["voice_clone_xtts_anna"]
-        or services_status["voice_clone_xtts_marina"]
-        or services_status["voice_clone_openvoice"]
-        or services_status["piper_tts"]
-    )
-    # In cloud mode, TTS is not required for healthy status
-    if DEPLOYMENT_MODE == "cloud":
-        core_ok = services_status["llm"]
-    else:
-        core_ok = any_tts and services_status["llm"]
-
-    # Get database status
-    db_status = await get_database_status()
-
-    result = {
-        "status": "healthy" if core_ok else "degraded",
-        "deployment_mode": DEPLOYMENT_MODE,
-        "services": services_status,
-        "database": db_status.get("database", {}),
-        "timestamp": datetime.now().isoformat(),
-    }
-
-    # Добавляем статистику streaming TTS если доступен
-    if streaming_tts_manager is not None:
-        result["streaming_tts_stats"] = streaming_tts_manager.get_stats()
-
-    # Internet monitor status
-    im = getattr(container, "internet_monitor", None)
-    if im is not None:
-        st = im.state
-        result["internet"] = {
-            "status": st.status.value,
-            "ping_ms": round(st.ping_ms, 1) if st.ping_ms else None,
-            "current_llm_backend": st.current_llm_backend,
-            "switch_count": st.switch_count,
-            "last_check": st.last_check,
-        }
-
-    return result
-
-
-@app.get("/admin/deployment-mode")
-async def get_deployment_mode():
-    """Return current deployment mode for frontend."""
-    return {"mode": DEPLOYMENT_MODE}
-
-
-def synthesize_with_current_voice(text: str, output_path: str, language: str = "ru"):
-    """
-    Синтезирует речь с текущим выбранным голосом.
-    Учитывает current_voice_config.
-
-    Engines:
-    - piper: CPU, быстрый, предобученные голоса (dmitri, irina)
-    - openvoice: GPU CC 6.1+, клонирование голоса (marina_openvoice)
-    - xtts: GPU CC >= 7.0, лучшее качество клонирования (anna, marina)
-    """
-    engine = current_voice_config["engine"]
-    voice = current_voice_config["voice"]
-
-    if engine == "piper" and piper_service:
-        logger.info(f"🎙️ Piper синтез ({voice}): '{text[:40]}...'")
-        piper_service.synthesize_to_file(text, output_path, voice=voice)
-    elif engine == "openvoice" and openvoice_service:
-        logger.info(f"🎙️ OpenVoice синтез (Марина): '{text[:40]}...'")
-        openvoice_service.synthesize_to_file(text, output_path, language=language)
-    elif engine == "xtts" and voice == "anna" and anna_voice_service:
-        logger.info(f"🎙️ XTTS синтез (Анна): '{text[:40]}...'")
-        anna_voice_service.synthesize_to_file(text, output_path, language=language)
-    elif engine == "xtts" and voice == "marina" and voice_service:
-        logger.info(f"🎙️ XTTS синтез (Марина): '{text[:40]}...'")
-        voice_service.synthesize_to_file(text, output_path, language=language)
-    elif anna_voice_service:
-        # Fallback to Анна if available (default)
-        logger.info(f"🎙️ XTTS синтез (Анна fallback): '{text[:40]}...'")
-        anna_voice_service.synthesize_to_file(text, output_path, language=language)
-    elif voice_service:
-        # Fallback to Марина if available
-        logger.info(f"🎙️ XTTS синтез (Марина fallback): '{text[:40]}...'")
-        voice_service.synthesize_to_file(text, output_path, language=language)
-    elif openvoice_service:
-        # Fallback to OpenVoice if XTTS not available
-        logger.info(f"🎙️ OpenVoice синтез (fallback): '{text[:40]}...'")
-        openvoice_service.synthesize_to_file(text, output_path, language=language)
-    elif piper_service:
-        # Fallback to Piper
-        logger.info(f"🎙️ Piper синтез (fallback): '{text[:40]}...'")
-        piper_service.synthesize_to_file(text, output_path, voice="irina")
-    else:
-        raise RuntimeError("No TTS service available")
-
-
-@app.post("/tts")
-async def text_to_speech(request: TTSRequest):
-    """
-    Синтез речи с текущим выбранным голосом
-    """
-    if not voice_service and not piper_service:
-        raise HTTPException(status_code=503, detail="No TTS service initialized")
-
-    try:
-        # Генерируем уникальное имя файла
-        output_file = TEMP_DIR / f"tts_{datetime.now().timestamp()}.wav"
-
-        # Синтезируем с текущим голосом
-        synthesize_with_current_voice(
-            text=request.text, output_path=str(output_file), language=request.language
-        )
-
-        # Возвращаем файл
-        return FileResponse(path=output_file, media_type="audio/wav", filename="response.wav")
-
-    except Exception as e:
-        logger.error(f"❌ TTS Error: {e}")
-        raise HTTPException(status_code=500, detail=str(e))
-
-
-@app.post("/stt")
-async def speech_to_text(audio: UploadFile = File(...)):
-    """
-    Распознавание речи из аудио файла
-    """
-    if not stt_service:
-        raise HTTPException(status_code=503, detail="STT service not initialized")
-
-    try:
-        # Сохраняем загруженный файл
-        temp_audio = TEMP_DIR / f"stt_{datetime.now().timestamp()}_{audio.filename}"
-
-        with open(temp_audio, "wb") as f:
-            content = await audio.read()
-            f.write(content)
-
-        # Распознаем
-        result = stt_service.transcribe(temp_audio, language="ru")
-
-        # Удаляем временный файл
-        temp_audio.unlink()
-
-        return {
-            "text": result["text"],
-            "language": result["language"],
-            "segments_count": len(result.get("segments", [])),
-        }
-
-    except Exception as e:
-        logger.error(f"❌ STT Error: {e}")
-        raise HTTPException(status_code=500, detail=str(e))
-
-
-@app.post("/chat")
-async def chat(request: ConversationRequest):
-    """
-    Получить ответ от LLM
-    """
-    if not llm_service:
-        raise HTTPException(status_code=503, detail="LLM service not initialized")
-
-    try:
-        response = llm_service.generate_response(request.text)
-
-        return {"response": response, "session_id": request.session_id}
-
-    except Exception as e:
-        logger.error(f"❌ LLM Error: {e}")
-        raise HTTPException(status_code=500, detail=str(e))
-
-
-@app.post("/process_call")
-async def process_call(audio: UploadFile = File(...)):
-    """
-    Полный цикл обработки звонка:
-    1. STT - распознавание речи
-    2. LLM - генерация ответа
-    3. TTS - синтез речи
-
-    Возвращает аудио с ответом секретаря
-    """
-    call_id = f"call_{datetime.now().strftime('%Y%m%d_%H%M%S')}"
-    logger.info(f"📞 Обработка звонка {call_id}")
-
-    try:
-        # 1. Сохраняем входящий аудио
-        input_audio = CALLS_LOG_DIR / f"{call_id}_input.wav"
-        with open(input_audio, "wb") as f:
-            content = await audio.read()
-            f.write(content)
-
-        # 2. Распознаем речь (STT)
-        logger.info(f"🎧 STT для {call_id}")
-        stt_result = stt_service.transcribe(input_audio, language="ru")
-        recognized_text = stt_result["text"]
-        logger.info(f"📝 Распознано: {recognized_text}")
-
-        # Сохраняем транскрипцию
-        with open(CALLS_LOG_DIR / f"{call_id}_transcript.txt", "w") as f:
-            f.write(f"USER: {recognized_text}\n")
-
-        # 3. Генерируем ответ (LLM)
-        logger.info(f"🤖 LLM для {call_id}")
-        llm_response = llm_service.generate_response(recognized_text)
-        logger.info(f"💬 Ответ: {llm_response}")
-
-        # Дополняем транскрипцию
-        with open(CALLS_LOG_DIR / f"{call_id}_transcript.txt", "a") as f:
-            f.write(f"ASSISTANT: {llm_response}\n")
-
-        # 4. Синтезируем ответ (TTS)
-        logger.info(f"🎙️  TTS для {call_id}")
-        output_audio = CALLS_LOG_DIR / f"{call_id}_output.wav"
-        voice_service.synthesize_to_file(
-            text=llm_response, output_path=str(output_audio), language="ru"
-        )
-
-        logger.info(f"✅ Звонок {call_id} обработан")
-
-        # 5. Возвращаем аудио ответ
-        return FileResponse(
-            path=output_audio,
-            media_type="audio/wav",
-            filename=f"{call_id}_response.wav",
-            headers={
-                "X-Call-ID": call_id,
-                "X-Recognized-Text": recognized_text,
-                "X-Response-Text": llm_response,
-            },
-        )
-
-    except Exception as e:
-        logger.error(f"❌ Ошибка обработки звонка {call_id}: {e}")
-        raise HTTPException(status_code=500, detail=str(e))
-
-
-@app.post("/reset_conversation")
-async def reset_conversation():
-    """Сброс истории диалога"""
-    if llm_service:
-        llm_service.reset_conversation()
-        return {"status": "ok", "message": "Conversation history reset"}
-    raise HTTPException(status_code=503, detail="LLM service not initialized")
-
-
-# ============== OpenAI-Compatible Endpoints for OpenWebUI ==============
-
-
-@app.get("/v1/models")
-@app.get("/v1/models/")
-async def list_models():
-    """OpenAI-compatible models list for OpenWebUI"""
-    # Определяем backend и суффикс для имени модели
-    if llm_service and hasattr(llm_service, "api_url"):
-        # vLLM backend - проверяем модель
-        model_name = getattr(llm_service, "model_name", "unknown")
-        if model_name == "lydia" or "qwen" in model_name.lower():
-            backend_suffix = "qwen"
-            backend_desc = "Qwen2.5-7B + LoRA"
-        elif "llama" in model_name.lower():
-            backend_suffix = "llama"
-            backend_desc = "Llama-3.1-8B"
-        else:
-            backend_suffix = "vllm"
-            backend_desc = model_name
-    elif llm_service and isinstance(llm_service, CloudLLMService):
-        backend_suffix = "cloud"
-        ptype = getattr(llm_service, "provider_type", "cloud")
-        backend_desc = f"{ptype}: {getattr(llm_service, 'model_name', 'unknown')}"
-    else:
-        backend_suffix = "cloud"
-        backend_desc = "Cloud AI"
-
-    return {
-        "object": "list",
-        "data": [
-            {
-                "id": f"anna-secretary-{backend_suffix}",
-                "object": "model",
-                "created": 1700000000,
-                "owned_by": "ai-secretary",
-                "permission": [],
-                "root": f"anna-secretary-{backend_suffix}",
-                "parent": None,
-                "description": f"Анна - цифровой секретарь ({backend_desc})",
-            },
-            {
-                "id": f"marina-secretary-{backend_suffix}",
-                "object": "model",
-                "created": 1700000000,
-                "owned_by": "ai-secretary",
-                "permission": [],
-                "root": f"marina-secretary-{backend_suffix}",
-                "parent": None,
-                "description": f"Марина - цифровой секретарь ({backend_desc})",
-            },
-        ],
-    }
-
-
-@app.get("/v1/voices")
-async def list_voices():
-    """List available voices"""
-    voices = []
-    if anna_voice_service:
-        voices.append({"voice_id": "anna", "name": "Анна", "language": "ru"})
-    if voice_service:
-        voices.append({"voice_id": "marina", "name": "Марина", "language": "ru"})
-    if piper_service:
-        voices.append({"voice_id": "dmitri", "name": "Дмитрий", "language": "ru"})
-        voices.append({"voice_id": "irina", "name": "Ирина", "language": "ru"})
-    return {"voices": voices}
-
-
-@app.post("/v1/audio/speech")
-async def openai_speech(request: OpenAISpeechRequest):
-    """
-    OpenAI-compatible TTS endpoint for OpenWebUI integration
-    POST /v1/audio/speech
-
-    Оптимизация: сначала проверяет кэш streaming TTS manager.
-    Если аудио уже было предсинтезировано во время streaming LLM - возвращает мгновенно.
-    """
-    if not voice_service and not piper_service:
-        raise HTTPException(status_code=503, detail="No TTS service initialized")
-
-    try:
-        output_file = TEMP_DIR / f"speech_{datetime.now().timestamp()}.wav"
-        start_time = time.time()
-
-        # Проверяем кэш streaming TTS (только для XTTS)
-        cached_audio = None
-        if current_voice_config["engine"] == "xtts" and streaming_tts_manager is not None:
-            cached_audio = streaming_tts_manager.get_cached_audio(request.input)
-
-        if cached_audio is not None:
-            # Cache HIT - используем предсинтезированное аудио
-            audio_data, sample_rate = cached_audio
-            sf.write(str(output_file), audio_data, sample_rate)
-            elapsed = time.time() - start_time
-            logger.info(f"⚡ TTS из кэша за {elapsed:.3f}s (vs ~5-10s обычный синтез)")
-        else:
-            # Cache MISS - синтезируем с текущим голосом
-            synthesize_with_current_voice(
-                text=request.input, output_path=str(output_file), language="ru"
-            )
-            elapsed = time.time() - start_time
-            logger.info(f"🎙️ TTS синтезирован за {elapsed:.2f}s")
-
-        return FileResponse(path=output_file, media_type="audio/wav", filename="speech.wav")
-
-    except Exception as e:
-        logger.error(f"❌ OpenAI TTS Error: {e}")
-        raise HTTPException(status_code=500, detail=str(e))
-
-
-@app.post("/v1/chat/completions")
-async def chat_completions(request: ChatCompletionRequest):
-    """
-    OpenAI-compatible chat completions endpoint for OpenWebUI
-    Supports both streaming and non-streaming responses.
-    При streaming - запускает фоновый синтез TTS по предложениям.
-    """
-    if not llm_service:
-        raise HTTPException(status_code=503, detail="LLM service not initialized")
-
-    logger.info(
-        f"💬 Chat completions request: stream={request.stream}, messages={len(request.messages)}"
-    )
-
-    # Конвертируем Pydantic модели в dict
-    messages = [{"role": m.role, "content": m.content} for m in request.messages]
-
-    if request.stream:
-        # Streaming response (SSE) с фоновым синтезом TTS
-        async def generate_stream():
-            created = int(time.time())
-            chunk_id = f"chatcmpl-{created}"
-            session_id = f"tts-{created}"
-
-            # Начинаем сессию streaming TTS если сервисы доступны
-            use_streaming_tts = streaming_tts_manager is not None and voice_service is not None
-
-            if use_streaming_tts:
-                streaming_tts_manager.start_session(session_id)
-                logger.info(f"🎬 Streaming TTS активирован для сессии {session_id}")
-
-            try:
-                for text_chunk in llm_service.generate_response_from_messages(
-                    messages, stream=True
-                ):
-                    # Отправляем chunk клиенту
-                    chunk_data = {
-                        "id": chunk_id,
-                        "object": "chat.completion.chunk",
-                        "created": created,
-                        "model": request.model,
-                        "choices": [
-                            {"index": 0, "delta": {"content": text_chunk}, "finish_reason": None}
-                        ],
-                    }
-                    yield f"data: {json.dumps(chunk_data, ensure_ascii=False)}\n\n"
-
-                    # Параллельно добавляем chunk в streaming TTS manager
-                    if use_streaming_tts and text_chunk:
-                        streaming_tts_manager.add_text_chunk(session_id, text_chunk, voice_service)
-
-                # Final chunk
-                final_chunk = {
-                    "id": chunk_id,
-                    "object": "chat.completion.chunk",
-                    "created": created,
-                    "model": request.model,
-                    "choices": [{"index": 0, "delta": {}, "finish_reason": "stop"}],
-                }
-                yield f"data: {json.dumps(final_chunk)}\n\n"
-                yield "data: [DONE]\n\n"
-
-                # Завершаем сессию TTS (склеивает и кэширует аудио)
-                if use_streaming_tts:
-                    # Запускаем в отдельном потоке чтобы не блокировать response
-                    threading.Thread(
-                        target=streaming_tts_manager.finish_session,
-                        args=(session_id, voice_service),
-                        daemon=True,
-                    ).start()
-
-            except Exception as e:
-                logger.error(f"❌ Streaming error: {e}")
-                error_chunk = {"error": {"message": str(e), "type": "server_error"}}
-                yield f"data: {json.dumps(error_chunk)}\n\n"
-
-        return StreamingResponse(
-            generate_stream(),
-            media_type="text/event-stream",
-            headers={
-                "Cache-Control": "no-cache",
-                "Connection": "keep-alive",
-            },
-        )
-    else:
-        # Non-streaming response
-        try:
-            response_text = llm_service.generate_response_from_messages(messages, stream=False)
-
-            # Consume generator if it returns one
-            if hasattr(response_text, "__iter__") and not isinstance(response_text, str):
-                response_text = "".join(response_text)
-
-            return {
-                "id": f"chatcmpl-{int(time.time())}",
-                "object": "chat.completion",
-                "created": int(time.time()),
-                "model": request.model,
-                "choices": [
-                    {
-                        "index": 0,
-                        "message": {"role": "assistant", "content": response_text},
-                        "finish_reason": "stop",
-                    }
-                ],
-                "usage": {"prompt_tokens": 0, "completion_tokens": 0, "total_tokens": 0},
-            }
-        except Exception as e:
-            logger.error(f"❌ Chat completions error: {e}")
-            raise HTTPException(status_code=500, detail=str(e))
-
-
 # ============== Admin Web Interface ==============
 # Vue 3 admin panel served from /admin (see Static Files section below)
 
 
 # ============== Admin API Endpoints ==============
-
-# ============== Auth Endpoints ==============
-
-
-@app.post("/admin/auth/login", response_model=LoginResponse)
-async def admin_login(request: LoginRequest, req: Request):
-    """
-    Authenticate user and return JWT token.
-
-    Default credentials: admin / admin
-    Set ADMIN_USERNAME and ADMIN_PASSWORD_HASH env vars for production.
-    """
-    user = await authenticate_user(request.username, request.password)
-    if not user:
-        raise HTTPException(status_code=401, detail="Invalid username or password")
-
-    response = await create_session(
-        username=user.username,
-        role=user.role,
-        user_id=user.id,
-        ip=req.client.host if req.client else None,
-        user_agent=req.headers.get("user-agent"),
-    )
-
-    # Audit log
-    await async_audit_logger.log(
-        action="login", resource="auth", user_id=user.username, details={"role": user.role}
-    )
-
-    return response
-
-
-@app.get("/admin/auth/me")
-async def admin_get_current_user(user: User = Depends(get_current_user)):
-    """Get current authenticated user info"""
-    return {"username": user.username, "role": user.role}
-
-
-@app.get("/admin/auth/status")
-async def admin_auth_status():
-    """Get authentication configuration status"""
-    return get_auth_status()
-
 
 # ============== Admin Models ==============
 
@@ -1683,81 +1089,6 @@ class AdminTTSTestRequest(BaseModel):
 
     text: str
     preset: str = "natural"
-
-
-@app.get("/admin/status")
-async def admin_status():
-    """Полный статус системы для админки"""
-    import torch
-
-    status = {
-        "orchestrator": "running",
-        "services": {
-            "voice_clone": voice_service is not None,
-            "llm": llm_service is not None,
-            "stt": stt_service is not None,
-            "streaming_tts": streaming_tts_manager is not None,
-            "piper_tts": piper_service is not None,
-        },
-        "gpu": None,
-        "streaming_tts_stats": None,
-        "llm_config": None,
-        "tts_config": None,
-    }
-
-    # GPU информация
-    if torch.cuda.is_available():
-        gpu_info = []
-        for i in range(torch.cuda.device_count()):
-            try:
-                name = torch.cuda.get_device_name(i)
-                total = torch.cuda.get_device_properties(i).total_memory / (1024**3)
-                allocated = torch.cuda.memory_allocated(i) / (1024**3)
-                gpu_info.append(
-                    {
-                        "id": i,
-                        "name": name,
-                        "total_gb": round(total, 2),
-                        "used_gb": round(allocated, 2),
-                    }
-                )
-            except Exception:
-                pass
-        status["gpu"] = gpu_info
-
-    # Streaming TTS статистика
-    if streaming_tts_manager:
-        status["streaming_tts_stats"] = streaming_tts_manager.get_stats()
-
-    # LLM конфигурация
-    if llm_service:
-        if hasattr(llm_service, "get_config"):
-            status["llm_config"] = llm_service.get_config()
-        else:
-            # Для vLLM и других сервисов без get_config
-            status["llm_config"] = {
-                "model_name": getattr(llm_service, "model_name", "unknown"),
-                "api_url": getattr(llm_service, "api_url", None),
-                "backend": (
-                    "vllm"
-                    if hasattr(llm_service, "api_url")
-                    else f"cloud:{getattr(llm_service, 'provider_id', 'unknown')}"
-                    if isinstance(llm_service, CloudLLMService)
-                    else "unknown"
-                ),
-            }
-
-    # TTS конфигурация
-    xtts_svc = anna_voice_service or voice_service
-    if xtts_svc:
-        status["tts_config"] = {
-            "device": xtts_svc.device,
-            "default_preset": xtts_svc.default_preset,
-            "samples_count": len(xtts_svc.voice_samples),
-            "cache_dir": str(xtts_svc.cache_dir),
-        }
-
-    return status
 
 
 @app.get("/admin/llm/prompt")
